fix: changed data location, improved GA handling
