+++ www/app/App.data.js
@@ -1,0 +1,11 @@
+import seoImage from '../shared/media/images/seo.png';
+
+const SEO_DATA = {
+    url: 'https://{project-domain}',
+    title: '{project-name}',
+    description: '{project-description}',
+    keywords: ['{project-keyword-1}', '{project-keyword-2}'],
+    image: { src: seoImage, width: 1200, height: 630 },
+};
+
+export default SEO_DATA;
