--- www/shared/data/index.js
@@ -1,2 +1,0 @@
-// Export Data
-export { default as contacts } from './contacts';
