+++ www/shared/styles/base/font-faces.css
@@ -1,8 +1,8 @@
-/* ==========================================================================
+/* ================================================================================
    Fonts
-   If you are going to include custom fonts please use the /src/media/fonts
+   If you are going to include custom fonts please use the /www/shared/media/fonts
    directory.
-   ========================================================================== */
+   ================================================================================ */
 
 @font-face {
     font-family: "Montserrat";
